+++ src/main/java/io/coodoo/appconfig/boundary/AppConfigs.java
@@ -19,7 +19,7 @@ import javax.persistence.EntityManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.coodoo.appconfig.boundary.annotation.AppConfigsEntityManager;
+import io.coodoo.appconfig.boundary.annotation.AppConfigEntityManager;
 import io.coodoo.appconfig.control.AppConfigSettings;
 import io.coodoo.appconfig.control.EncryptDecrypt;
 import io.coodoo.appconfig.entity.AppConfigValue;
@@ -35,7 +35,7 @@ public class AppConfigs {
     private static Logger log = LoggerFactory.getLogger(AppConfigs.class);
 
     @Inject
-    @AppConfigsEntityManager
+    @AppConfigEntityManager
     EntityManager entityManager;
 
     public String getString(AppConfigKey key) {